+++ packages/wouter/lib/src/models/models.dart
@@ -3,3 +3,5 @@ export 'regexp_data/model.dart';
 export 'route_entry/model.dart';
 export 'state/state.dart';
 export 'actions.dart';
+export 'matchers/matchers.dart';
+export 'match_data/model.dart';