+++ packages/wouter/lib/src/navigators/models/models.dart
@@ -1,2 +1,0 @@
-export 'functions.dart';
-export 'match_data/model.dart';