+++ packages/wouter/lib/src/widgets/navigators/navigators.dart
@@ -3,5 +3,3 @@ export 'base.dart';
 export 'tab.dart';
 export 'listenable.dart';
 export 'page.dart';
-export 'models/models.dart';
-export 'matchers/matchers.dart';